added task 114 from Foncode

+++ Foncode68.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 
-//*
+/*
 int main() {
-	int inp; std::cin >> inp;
+	long inp; std::cin >> inp;
 
-	for (int i = 2; inp/i >= i;) {
+	for (long i = 2; inp/i >= i;) {
 		if (inp % i != 0) { i++; continue; }
 		std::cout << i << ' ';
 		inp /= i;
